+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## 1.4.0 (not yet released)
 
-- MAINT: Removing 'DATABASE_URL' environment-var from entrypoint (#1037)
+- ENH: Added select filter options: '(shared|published) by you' (#750)
 - ENH: Batch upload (#173, #877, #906)
 - ENH: Button to regenerate cached properties (thumbnail, DZI, etc., #895)
 - ENH: Added creation and last modified datetime to surfaces and topographies (#1010)
 - BUG: Fixes related to uploading files (#223, #261, #667)
+- MAINT: Removing 'DATABASE_URL' environment-var from entrypoint (#1037)
 - MAINT: Surface view is now largely a single-page application
 - MAINT: REST API for topography handling (#173)
 - MAINT: Major refactor of task handling code